improve error handling for universal delete

+++ versions/V0.2.1/database/services/delete.py
@@ -1,4 +1,6 @@
 from options import args
+from psycopg2.errors import UndefinedTable
+from database.repositories.factory import RepositoryNotFoundError
 from database.repositories.generic import GenericRepository
 from database.repositories.shareholder import ShareholderRepository
 
@@ -8,10 +10,19 @@ logger = logging.getLogger(__name__)
 
 def handle_delete_by_id(db):
     try:
+        if not args.table:
+            logger.error('Table name not provided.')
+            print('Error: Table name is required for deletion')
+            return
+        
         repository = GenericRepository(db, args.table)
         repository.delete(args.Remove)
         logger.info(f'Entity with id: {args.Remove} from table: {args.table} deleted sucessfully.')
-        
+    
+    except RepositoryNotFoundError as e:
+        logger.error(f'Repository not found error: {e}')
+    except UndefinedTable as e:
+        logger.error(f'Table not found, deletion failed: {e}')
     except Exception as e:
         logger.error(f'An error occurred handling the deletion of an entity from the table: {e}')
         raise
